# docs/swagger.yaml
definitions:
  attachment.Attachment:
    properties:
      createdAt:
        type: string
      height:
        type: integer
      id:
        type: string
      mime:
        type: string
      serverId:
        type: string
      size:
        type: integer
      thumbUrl:
        type: string
      url:
        type: string
      userId:
        type: string
      width:
        type: integer
    type: object
  channel.ChannelInfo:
    properties:
      id:
        type: string
      name:
        type: string
      serverId:
        type: string
      type:
        type: string
    type: object
  channel.Member:
    properties:
      avatarId:
        type: string
      deafened:
        type: boolean
      displayName:
        type: string
      id:
        type: string
      muted:
        type: boolean
    type: object
  handlers.ChannelDetailResponse:
    properties:
      connected:
        items:
          $ref: '#/definitions/channel.Member'
        type: array
      connectedCount:
        type: integer
      id:
        type: string
      name:
        type: string
      serverId:
        type: string
      type:
        type: string
    type: object
  handlers.InviteCreatedResponse:
    properties:
      inviteToken:
        type: string
      url:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      altcha:
        example: eyJhbGdvcml0aG0iOi...
        type: string
      password:
        example: s3cr3t-p4ss
        type: string
      username:
        example: admin
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        example: eyJhbG...
        type: string
    type: object
  handlers.MeResponse:
    properties:
      avatarId:
        type: string
      avatarUrl:
        type: string
      displayName:
        type: string
      id:
        type: string
    type: object
...
  message.Message:
    properties:
      attachmentId:
        type: string
      channelId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      serverId:
        type: string
      userId:
        type: string
    type: object
  pow.Challenge:
    properties:
      algorithm:
        type: string
      challenge:
        type: string
      maxnumber:
        type: integer
      salt:
        type: string
      signature:
        type: string
    type: object
  server.MemberInfo:
    properties:
      avatarId:
        type: string
      avatarUrl:
        type: string
      displayName:
        type: string
      id:
        type: string
      isOwner:
        type: boolean
      online:
        type: boolean
    type: object
  server.ServerInfo:
    properties:
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
    type: object
info:
  contact: {}
  description: Armonic API backend server.
  title: Armonic API
  version: "1.0"
paths:
  /attachment/{id}:
    get:
      description: Member-only. Serves the sanitized image bytes.
      parameters:
      - description: Attachment ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: attachment not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Fetch an attachment
      tags:
      - Attachment
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user in already claimed server
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: invalid body
          schema:
            type: string
        "401":
          description: invalid credentials
          schema:
            type: string
        "403":
          description: server not claimed yet
          schema:
            type: string
        "409":
          description: proof of work expired or already used
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      summary: Login user
      tags:
      - Users
  /channel/{id}:
    get:
      consumes:
      - application/json
      description: Get a single channel by its id, enriched with live voice state
        (who is connected right now) when it's a voice channel
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.ChannelDetailResponse'
        "400":
          description: invalid channel id
          schema:
            type: string
        "404":
          description: channel not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Channel info by id
      tags:
      - Channel
  /channel/{id}/messages:
    get:
      description: Get a page of messages for a channel (newest first)
      parameters:
      - description: Channel ID
        in: path
        name: id
        required: true
        type: string
      - description: Max messages to return (default 50, max 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/message.Message'
            type: array
        "400":
          description: invalid channel id
          schema:
            type: string
        "404":
          description: channel not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Channel messages
      tags:
      - Channel
...
  /me:
    get:
      description: The authenticated caller's id, display name and avatar.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MeResponse'
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Own profile
      tags:
      - Users
  /me/avatar:
    post:
      consumes:
      - multipart/form-data
      description: Uploads an image through the same pipeline as attachments and points
        the caller's avatar at it.
      parameters:
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/attachment.Attachment'
        "413":
          description: file exceeds the maximum upload size
          schema:
            type: string
        "415":
          description: unsupported image format
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Set own avatar
      tags:
      - Users
  /pow/challenge:
    get:
      description: Issues an Altcha-compatible challenge. Solve it and send the base64
        solution in the "altcha" field of the login, claim or invite-signup request.
        Returns 404 when POW_ENABLED is off.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pow.Challenge'
        "404":
          description: proof of work is disabled
          schema:
            type: string
      summary: Proof-of-work challenge
      tags:
      - Users
  /server:
    get:
      description: List the servers the authenticated user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.ServerInfo'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: My servers
      tags:
      - Server
  /server/{id}:
    get:
      consumes:
      - application/json
      description: Get channels from server with server-id
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/channel.ChannelInfo'
            type: array
      security:
      - BearerAuth: []
      summary: Server id info
      tags:
      - Server
  /server/{id}/invite:
    post:
      description: Owner-only. Mint a single-use invite link for a server.
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.InviteCreatedResponse'
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create invite
      tags:
      - Invite
  /server/{id}/members:
    get:
      description: 'Member-only. The roster of a server: every member with their display
        name, avatar and owner flag, plus whether they currently have a live WebSocket
        connection.'
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.MemberInfo'
            type: array
        "400":
          description: invalid server id
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Server members
      tags:
      - Server
  /server/{id}/upload:
    post:
      consumes:
      - multipart/form-data
      description: Member-only. Multipart upload of one image (field "file"). The
        file is validated by magic number, decoded, stripped of all metadata by re-encoding
        its pixels, thumbnailed and stored. The uploaded bytes are discarded.
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      - description: Image file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/attachment.Attachment'
        "400":
          description: image could not be decoded
          schema:
            type: string
        "413":
          description: file exceeds the maximum upload size
          schema:
            type: string
        "415":
          description: unsupported image format
          schema:
            type: string
        "422":
          description: image dimensions exceed the configured limit
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upload an image
      tags:
      - Attachment
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT returned by /auth/login.
      Example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...